feat: add image asset processor service for HEIC and image processing
- Implemented ImageAssetProcessorService to handle HEIC image conversion and processing.
- Added utility types for file assets including LocalFileSignature and HeicConvert.
- Created FlowExecutionHistoryService to manage flow execution history with methods for starting, finalizing, and updating progress.
- Added tests for FileAssetAccessService and FlowExecutionHistoryService to ensure functionality and cache invalidation.
