+++ src/modules/file-management/index.ts
@@ -1,5 +1,8 @@
 export * from './services/file-assets.service';
+export * from './services/file-asset-access.service';
+export * from './services/image-asset-processor.service';
 export * from './services/file-management.service';
+export * from './types/file-asset.types';
 export * from './storage/gcs-storage.service';
 export * from './storage/local-storage.service';
 export * from './storage/r2-storage.service';
