+++ src/modules/flow/types/flow-execution-history.types.ts
@@ -1,0 +1,8 @@
+export type FlowExecutionHistoryId = number | string | null;
+
+export type FlowProgressSnapshot = {
+  completedSteps?: any[];
+  currentStep?: string | null;
+  failedStep?: string | null;
+  totalSteps?: number;
+};
